Unify network interface config for AWS and OpenStack providers at the network ID level.
When launching instances with private networking, AWS requires a subnet ID while OpenStack requires a network ID. Until now, this distinction had to be made at the application level vs. within automatically CloudBridge. With this change, the network ID needs to be supplied for either provider. For AWS, CloudBridge will extract a subnet within the given network and use it. This clearly leads to less control for the application/user but it provides a uniform interface. We can revisit this implementation  after we see how it works in practice. Most likely, we will want to check for multiple subnets within a network and perhaps choose one with the largest available address range.