Update the Getting Started tutorial with instructions for using private networking.
Note that there is still a problem here with OpenStack: OS requires a router to be attached to an external network (i.e., a gateway) vs. a user-created network. We do not have a way to filter for an external network since this is only an OS concept. So the example provided will require an extra step of listing all the available networks, picking out the public/external one and attaching the router to it. This also happens to be a problem for writing tests, hence those are still missing.