Version Conflict when trying to run tox tests #37
If you clone the repository, setup the needed environment and then run
tox, it fails with a version conflict error.

This happens because the dependencies in setup.py are best resolved by
dropping down to and installing requests 2.12.5. However, that version
of requests in turn explicitly states that it requires Babel==2.3.4 and
not the version Babel==2.4.0 that has already been installed.

Instant end of run :(

This commit makes the version of Babel explicit