OpenStackFloatingIP in_use method doesn't return the correct result #42
The floating IP status was tests, and if found to be 'ACTIVE' True was
returned. However, on our nectar cloud a newly created floating IP is
set to 'ACTIVE' by default - even though it is not in use.

This commit uses port id as indicator of being in use: if set, then
the floating ip is attached, if not, it is not.