Deal with creating a subnet with overlapping range
The GCE provider creates auto-mode new networks. So subnets are created
automatically. So, it is possible that when a test creates a custom subnet, even
if it is right after creating a new network, the provided CIDR block overlap
with an existing subnet. In those cases, we just return the existing subnet.

Test expectations are modified accordingly.