Removed incorrect object identity test
Currently, this line tests object equality. However, this is incorrect
because the object equality test takes into account transient properties
like state, and therefore, this test fails if a state transition occurs.
Instead, this should simply check for identity only, and not equality,
which is done in the subsequent line.