Fix creating instances with VM type and image IDs
It works with GCEVMType and GCEMachineImage, but not when a VM type ID or an
image ID is given.

Also, make images service get() method search in public images, too, when the
given image name does not exist in the project.

Test stats after this CL
========================
Ran 58 tests in 1402.766s

FAILED (SKIP=1, errors=26, failures=8)